refactor(icons): rename svg to mdi-svg
