+++ packages/vuetify/src/services/icons/index.ts
@@ -10,7 +10,7 @@ import presets from './presets'
 export class Icons extends Service {
   static property = 'icons'
 
-  public iconfont: VuetifyIconOptions['iconfont'] = 'svg'
+  public iconfont: VuetifyIconOptions['iconfont'] = 'mdi-svg'
   public values: VuetifyIconOptions['values'] = presets[this.iconfont]
 
   constructor (options: Partial<VuetifyIconOptions> = {}) {
