+++ packages/vuetify/src/services/icons/presets/index.ts
@@ -1,12 +1,16 @@
 // Types
 import { VuetifyIconSets } from 'vuetify/types/services/icons'
 
-import svg from './svg'
+import mdiSvg from './mdi-svg'
 import md from './md'
 import mdi from './mdi'
 import fa from './fa'
 import fa4 from './fa4'
 
 export default Object.freeze({
-  svg, md, mdi, fa, fa4
+  'mdi-svg': mdiSvg,
+  md,
+  mdi,
+  fa,
+  fa4
 }) as VuetifyIconSets
