+++ packages/vuetify/src/services/icons/presets/mdi-svg.ts
